adjustment for new module loader
Recent changes in module loading require specifying "module: elasticsearch". In the past, it was possible that the plugin got loaded twice when the name "elasticsearch" was used. The output config name was changed to "logsene" and we added the module name. In some cases, it would be useful to specify additional command line arguments e.g. via env variable LA_ARGUMENTS